Poprawienie Registry, rozpoczęcie implementacji strony szczegółowej danego kryminalisty
